Add ConstrainedBackTrackingSolver
This solver does an initial few passes (exact count depends on the
puzzle). Each pass keeps track of the values a cell can take and stores
them in the cell. This reduces the search space by elliminating any
values that would be obviously wrong based on the fixed cells.

If any cell only has 1 possible value, it is fixed. The rest of the
implementation is similar to the simple backtracker, with the only
change being that the cell increment skips any values that are invalid.